requirements.txt
numpy==1.19.5
opencv_python==4.5.1.48
pandas==1.2.1
matplotlib==3.4.1
Pillow==8.2.0
scikit_learn==0.24.2
seaborn==0.11.1
